Also test the latest published machine with challenge tests.
This helps to ensure that the machine works with one-step-proofs.
