+++ system_tests/block_validator_test.go
@@ -9,10 +9,7 @@ package arbtest
 
 import (
 	"context"
-	"io"
 	"math/big"
-	"net/http"
-	"os"
 	"testing"
 	"time"
 
@@ -262,34 +259,6 @@ func testBlockValidatorSimple(t *testing.T, opts Options) {
 	if finalRefCount < 0 || finalRefCount > int64(largestRefCount) {
 		Fatal(t, "unexpected refcount:", finalRefCount)
 	}
-}
-
-func populateMachineDir(t *testing.T, cr *github.ConsensusRelease) string {
-	baseDir := t.TempDir()
-	machineDir := baseDir + "/machines"
-	err := os.Mkdir(machineDir, 0755)
-	Require(t, err)
-	err = os.Mkdir(machineDir+"/latest", 0755)
-	Require(t, err)
-	mrFile, err := os.Create(machineDir + "/latest/module-root.txt")
-	Require(t, err)
-	_, err = mrFile.WriteString(cr.WavmModuleRoot)
-	Require(t, err)
-	machResp, err := http.Get(cr.MachineWavmURL.String())
-	Require(t, err)
-	defer machResp.Body.Close()
-	machineFile, err := os.Create(machineDir + "/latest/machine.wavm.br")
-	Require(t, err)
-	_, err = io.Copy(machineFile, machResp.Body)
-	Require(t, err)
-	replayResp, err := http.Get(cr.ReplayWasmURL.String())
-	Require(t, err)
-	defer replayResp.Body.Close()
-	replayFile, err := os.Create(machineDir + "/latest/replay.wasm")
-	Require(t, err)
-	_, err = io.Copy(replayFile, replayResp.Body)
-	Require(t, err)
-	return machineDir
 }
 
 func TestBlockValidatorSimpleOnchainUpgradeArbOs(t *testing.T) {
