+++ system_tests/full_challenge_mock_test.go
@@ -8,14 +8,16 @@ import "testing"
 
 func TestMockChallengeManagerAsserterIncorrect(t *testing.T) {
 	t.Parallel()
+	defaultWasmRootDir := ""
 	for i := int64(1); i <= makeBatch_MsgsPerBatch*3; i++ {
-		RunChallengeTest(t, false, true, i)
+		RunChallengeTest(t, false, true, i, defaultWasmRootDir)
 	}
 }
 
 func TestMockChallengeManagerAsserterCorrect(t *testing.T) {
 	t.Parallel()
+	defaultWasmRootDir := ""
 	for i := int64(1); i <= makeBatch_MsgsPerBatch*3; i++ {
-		RunChallengeTest(t, true, true, i)
+		RunChallengeTest(t, true, true, i, defaultWasmRootDir)
 	}
 }
