+++ system_tests/full_challenge_test.go
@@ -6,14 +6,37 @@
 
 package arbtest
 
-import "testing"
+import (
+	"context"
+	"testing"
+
+	"github.com/offchainlabs/nitro/util/testhelpers/github"
+)
 
 func TestChallengeManagerFullAsserterIncorrect(t *testing.T) {
 	t.Parallel()
-	RunChallengeTest(t, false, false, makeBatch_MsgsPerBatch+1)
+	defaultWasmRootDir := ""
+	RunChallengeTest(t, false, false, makeBatch_MsgsPerBatch+1, defaultWasmRootDir)
+}
+
+func TestChallengeManagerFullAsserterIncorrectWithPublishedMachine(t *testing.T) {
+	t.Parallel()
+	cr, err := github.LatestConsensusRelease(context.Background())
+	Require(t, err)
+	machPath := populateMachineDir(t, cr)
+	RunChallengeTest(t, false, true, makeBatch_MsgsPerBatch+1, machPath)
 }
 
 func TestChallengeManagerFullAsserterCorrect(t *testing.T) {
 	t.Parallel()
-	RunChallengeTest(t, true, false, makeBatch_MsgsPerBatch+2)
+	defaultWasmRootDir := ""
+	RunChallengeTest(t, true, false, makeBatch_MsgsPerBatch+2, defaultWasmRootDir)
+}
+
+func TestChallengeManagerFullAsserterCorrectWithPublishedMachine(t *testing.T) {
+	t.Parallel()
+	cr, err := github.LatestConsensusRelease(context.Background())
+	Require(t, err)
+	machPath := populateMachineDir(t, cr)
+	RunChallengeTest(t, true, true, makeBatch_MsgsPerBatch+2, machPath)
 }
